chiselc: Move helper functions to transforms/utils.rs
Preparation for using then for `findMany` inference.
