+++ chiselc/src/transforms/filter.rs
@@ -1,20 +1,14 @@
 use crate::filtering::FilterProperties;
-use crate::query::BinaryExpr as QBinaryExpr;
-use crate::query::BinaryOp as QBinaryOp;
 use crate::query::Expr as QExpr;
 use crate::query::Filter as QFilter;
 use crate::query::Literal as QLiteral;
 use crate::query::Operator as QOperator;
-use crate::query::PropertyAccessExpr as QPropertyAccessExpr;
 use crate::query::Scan as QScan;
 use crate::symbols::Symbols;
+use crate::transforms::utils::{convert_bin_expr, lookup_callee_entity_type};
 use crate::utils::{is_call_to_entity_method, is_ident_member_prop, pat_to_string};
-use anyhow::{anyhow, Result};
 
-use swc_ecmascript::ast::{
-    BinExpr, BinaryOp, BlockStmtOrExpr, CallExpr, Callee, Expr, Ident, Lit, MemberExpr, MemberProp,
-    Stmt,
-};
+use swc_ecmascript::ast::{BlockStmtOrExpr, CallExpr, Callee, Expr, Lit, Stmt};
 
 /// Infer filter operator from the lambda predicate of to filter()
 pub fn infer_filter(
@@ -109,71 +103,4 @@ fn is_rewritable_filter(callee: &Callee, symbols: &Symbols) -> bool {
         },
         _ => false,
     }
-}
-
-fn lookup_callee_entity_type(callee: &Callee) -> Result<String> {
-    match callee {
-        Callee::Expr(expr) => lookup_entity_type(expr),
-        _ => {
-            todo!();
-        }
-    }
-}
-
-fn lookup_entity_type(expr: &Expr) -> Result<String> {
-    match expr {
-        Expr::Member(MemberExpr { obj, .. }) => lookup_entity_type(obj),
-        Expr::Call(CallExpr { callee, .. }) => lookup_callee_entity_type(callee),
-        Expr::Ident(Ident { sym, .. }) => Ok(sym.to_string()),
-        _ => {
-            anyhow::bail!("Failed to look up entity type from call chain")
-        }
-    }
-}
-
-fn convert_bin_expr(expr: &BinExpr) -> Result<QExpr> {
-    let left = Box::new(convert_expr(&expr.left)?);
-    let op = convert_binary_op(&expr.op)?;
-    let right = Box::new(convert_expr(&expr.right)?);
-    Ok(QExpr::BinaryExpr(QBinaryExpr { left, op, right }))
-}
-
-fn convert_expr(expr: &Expr) -> Result<QExpr> {
-    match expr {
-        Expr::Bin(bin_expr) => convert_bin_expr(bin_expr),
-        Expr::Paren(paren_expr) => Ok(convert_expr(&*paren_expr.expr)?),
-        Expr::Lit(Lit::Num(number)) => Ok(QExpr::Literal(QLiteral::Num(number.value))),
-        Expr::Lit(Lit::Str(s)) => Ok(QExpr::Literal(QLiteral::Str(format!("{}", s.value)))),
-        Expr::Member(member_expr) => {
-            let obj = convert_expr(&member_expr.obj)?;
-            let prop = match &member_expr.prop {
-                MemberProp::Ident(ident) => ident.sym.to_string(),
-                _ => {
-                    todo!();
-                }
-            };
-            Ok(QExpr::PropertyAccess(QPropertyAccessExpr {
-                object: Box::new(obj),
-                property: prop,
-            }))
-        }
-        Expr::Ident(ident) => Ok(QExpr::Identifier(ident.sym.to_string())),
-        _ => Err(anyhow!("Unsupported expression: {:#?}", expr)),
-    }
-}
-
-fn convert_binary_op(op: &BinaryOp) -> Result<QBinaryOp> {
-    Ok(match op {
-        BinaryOp::EqEq => QBinaryOp::Eq,
-        BinaryOp::Gt => QBinaryOp::Gt,
-        BinaryOp::GtEq => QBinaryOp::GtEq,
-        BinaryOp::Lt => QBinaryOp::Lt,
-        BinaryOp::LtEq => QBinaryOp::LtEq,
-        BinaryOp::NotEq => QBinaryOp::NotEq,
-        BinaryOp::LogicalAnd => QBinaryOp::And,
-        BinaryOp::LogicalOr => QBinaryOp::Or,
-        _ => {
-            anyhow::bail!("Cannot convert binary operator {}", op);
-        }
-    })
 }
