+++ chiselc/src/transforms/mod.rs
@@ -1,1 +1,2 @@
 pub mod filter;
+pub mod utils;
